BCO-4660 Schema Changes

Closes BCO-4660

See merge request aboutyou/cc-babo/admin-api!190

+++ src/main/java/com/aboutyou/cloud/adminapi/model/OrderReductionAmount.java
@@ -21,6 +21,6 @@ public class OrderReductionAmount implements ApiObjectInterface {
     * 
     */
     @SerializedName("relative")
-    Integer relative;
+    Double relative;
 
 }
